+++ packages/core/src/lib/utils.js
@@ -2,6 +2,10 @@ export function assign(...args) {
     return Object.assign(...args);
 }
 
+export function keys(object) {
+    return Object.keys(object);
+}
+
 export function floor(value) {
     return Math.floor(value);
 }
@@ -12,10 +16,6 @@ export function min(...args) {
 
 export function max(...args) {
     return Math.max(...args);
-}
-
-export function isObject(test) {
-    return typeof test === 'object' && test !== null;
 }
 
 export function symbol() {